split metrics tests
