+++ python/ray/serve/_private/test_utils.py
@@ -18,6 +18,10 @@ from ray import serve
 from ray._common.network_utils import build_address
 from ray._common.test_utils import wait_for_condition
 from ray._common.utils import TimerBase
+from ray._private.test_utils import (
+    PrometheusTimeseries,
+    fetch_prometheus_metric_timeseries,
+)
 from ray.actor import ActorHandle
 from ray.serve._private.client import ServeControllerClient
 from ray.serve._private.common import (
@@ -862,3 +866,94 @@ def request_with_retries(timeout=30, app_name=SERVE_DEFAULT_APP_NAME):
     except RuntimeError as e:
         # Preserve previous API by raising TimeoutError on expiry
         raise TimeoutError from e
+
+
+# Metrics test utilities
+TEST_METRICS_EXPORT_PORT = 9999
+
+
+def get_metric_float(
+    metric: str,
+    expected_tags: Optional[Dict[str, str]],
+    timeseries: Optional[PrometheusTimeseries] = None,
+) -> float:
+    """Gets the float value of metric.
+
+    If tags is specified, searched for metric with matching tags.
+
+    Returns -1 if the metric isn't available.
+    """
+    if timeseries is None:
+        timeseries = PrometheusTimeseries()
+    samples = fetch_prometheus_metric_timeseries(
+        [f"localhost:{TEST_METRICS_EXPORT_PORT}"], timeseries
+    ).get(metric, [])
+    for sample in samples:
+        if expected_tags.items() <= sample.labels.items():
+            return sample.value
+    return -1
+
+
+def check_metric_float_eq(
+    metric: str,
+    expected: float,
+    expected_tags: Optional[Dict[str, str]],
+    timeseries: Optional[PrometheusTimeseries] = None,
+) -> bool:
+    """Check if a metric's float value equals the expected value."""
+    metric_value = get_metric_float(metric, expected_tags, timeseries)
+    assert float(metric_value) == expected
+    return True
+
+
+def get_metric_dictionaries(
+    name: str, timeout: float = 20, timeseries: Optional[PrometheusTimeseries] = None
+) -> List[Dict]:
+    """Gets a list of metric's tags from metrics' text output.
+
+    Return:
+        Example:
+
+        >>> get_metric_dictionaries("ray_serve_num_http_requests")
+        [
+            {
+                'Component': 'core_worker',
+                'JobId': '01000000',
+                ...
+                'method': 'GET',
+                'route': '/hello'
+            },
+            {
+                'Component': 'core_worker',
+                ...
+                'method': 'GET',
+                'route': '/hello/'
+            }
+        ]
+    """
+    if timeseries is None:
+        timeseries = PrometheusTimeseries()
+
+    def metric_available() -> bool:
+        assert name in fetch_prometheus_metric_timeseries(
+            [f"localhost:{TEST_METRICS_EXPORT_PORT}"], timeseries
+        )
+        return True
+
+    wait_for_condition(metric_available, retry_interval_ms=1000, timeout=timeout)
+    serve_samples = [
+        sample
+        for sample in timeseries.metric_samples.values()
+        if "ray_serve_" in sample.name
+    ]
+    print(
+        "metrics", "\n".join([f"Labels: {sample.labels}\n" for sample in serve_samples])
+    )
+
+    metric_dicts = []
+    for sample in timeseries.metric_samples.values():
+        if sample.name == name:
+            metric_dicts.append(sample.labels)
+
+    print(metric_dicts)
+    return metric_dicts
