+++ python/ray/serve/tests/BUILD.bazel
@@ -249,6 +249,7 @@ py_test_module_list(
         "test_deploy_app_2.py",
         "test_metrics.py",
         "test_metrics_2.py",
+        "test_metrics_3.py",
     ],
     tags = [
         "exclusive",
